Object Destructuring: rm confusing getASTNode() fn
It's not entirely claer what getASTNode() returns.
Is this a single node? Many unfamiliar with ASTs
might assume that it's an `{op: ..., lhs: ...,
rhs: ...}` which maps to this example, but in the
subsequent exaple the node contains another node
which isn't immediately evident and which makes
the example more complex to grasp.